Create the settings spreadsheet programmatically.
This gets rid of the need for the go/ link and for users to ever have to copy-paste
spreadsheet URLs. Also, use the spreadsheet name instead of ID for looking it up.
That way it only needs to be generated once and will work across devices.
Only downside is that it breaks if a user changes the name of the sheet, but that
seems like the least of the concerns with using a spreadsheet as the backend.