Switch back to using threads.modify.
See the comments in the patch. messages.batchModify is really buggy on some
threads. Some threads work fine, but others are really problematic. And
sometimes it's just the INBOX label and other times it's user labels.
For now, swithc back to threads.modify and try to minimized the likelihood
of a race by checking if a thread has new messages right before modifying
and bailing if it does.