Rework how ThreadListView and ThreadListModel store state.
Make ThreadListView operate almost entirely on the DOM and make
ThreadListModel operate entirely on Threads. This leads to a cleaner
separation and simplifies a bunch of code.

- Make ThreadListModel hold a flat, sorted list of Threads instead of trying
to group them by queue.
- Make ThreadListView do all it's focus navigation based off the DOM
instead of the model.
- Change the pointers to the rendered thread and focused thread in
ThreadListView to be pointers to DOM nodes.
- Change Thread to always have message details once it's been fetched
and change the models to proactively fetch the message details as one
more incremental step to being able to assume the message details are
already there. As a result, ThreadListView and most of ThreadListModel
can call sync getters on Thread, which allows us to make a bunch of the
rendering logic sync and thus minimize race conditions.
- Change TodoModel and TriageModel to first add all the threads that
have message details on disk atomically to minimize user visible
inconsistencies, then incrementally process all the threads that need to
hit the network.
- Render on frame boundaries instead of synchronously. This gives a
natural throttling for cases where rendering takes a long time.
- Merge fetchThreads calls in TriageModel so we do a single fetchThreads
call, which mininizes round trips in the common cases where all the
threads have been processed.