Clear listeners on the model.
Now that the model is long-lived, we need to clear the listeners on the
model. Otherwise, the model keeps the ThreadListView alive and
listening to updates on the model and everything goes haywire. In
particular, this would lead to a bug where archiving a thread would take
you back to the threadlist instead of the next thread in single thread
mode.