Allow for changing the post send action in quick reply.
In addition to archive, you should be able to give it into an appropriate
priority or the blocked queue so that you can followup if needed.

As part of this, switch quick reply to using a RadialProgress instead of
HTMLProgressElement and give RadialProgress a concept of overflow where
it turns red.

Fixes issue #18.