Fetch all labels in TodoModel and TriageModel.
Gmail will sometimes return 0 thread count for labels to the labels.get
API. https://issuetracker.google.com/issues/36759403

Now that we do all the label fetches in a single fetchThreads call we
don't actually need to get the count for labels anymore and can just do
the fetch for all labels. This also turns out to be faster because it
avoids the extra round trip for the labels.get call!