Don't show toolbar or do updates when compose view has query parameters.
This keeps the prefilled compose view just for quick composes, which is
especially important for having homescreen shortcuts on android where it
will keep opening the app to the previous location you had it open if it
hasn't been killed. So navigating to triage and then going to the
homescreen button doesn't take you back to compose the way you'd want
to.