Make ThreadData a real class.
This way we can change Thread to always have message details and rely on
the type system to ensure that we have message details whenever we have
an instance of a Thread. A followup will be to make all the getters on
Thread sync now that they must have message details.