Really make it impossible for Threads to have uninitialized getters.
Have Threads take the message data as a constructor argument. They can
still update that data later, but this way we ensure that if we have a
Thread, then it definitely has message data. We can't just do all this
work in the constructor unfortunately because some of the work is async.