Fix race conditions in view creation.
If we switch to another view while we're in the process of creating the
current view, then we need to abort view creation. This was causing bugs
like rerendering the triage view when you switch from it to the compose
view, but main.js would think you were still in the compose view.