Check focused rows by default.
This makes it more clear what will happen when you click a button in the
toolbar since something is actually checked. If nothing is checked the
toolbar will no longer act on the focused row, which is a state you can
only get in by focusing a row and then clicking to uncheck it.