Only dispatch storage update events for some keys.
We don't want to reset all the models/views if something like
last_dequeue_time is updated. We really only want to do it when settings
are changed. Also, we don't want to dispatch this event the first time
we get the snapshot data. For the compose view, this would cause it to
reload once the firebase login finished.