Add regexp filter rules and change address filters to use them
Address filters currently do weird bespoke things. Make them just be regular filters
that try to substring or regexp match the name and/or the address in the to/cc/bcc/from
headers. We still parse the addresses first. This matches gmail behavior where you can't
do a substring search the spans both the name and the address.