Fix processing of new messages.
We don't need the labelId and priorityId guards anymore now that we trigger
based off the lack of a mktime label on a message. If there's any message
in the thread without a mktime label, then it needs to have filters applied.
This fixes the issue where threads would not get refiltered if they were
already triaged and a new message came in.