Fix boolean logic errors in allowViewMessages setting.
- We'd create a new ThreadListView on view transitions, but only set
allowViewMessages in handleThreadListChanged_. Instead, always set
allowViewMessages_ from the model.
- Fix Skim view to get the right default value for allowViewMessages and
make the button for entering/exiting triage be a toggle instead of a one
time thing now that the data is stored in the model and so will survive
view transitions.
- Sneak in a change to autosend to clear the form fields. autosend
is not a case where you want to repeatedly fillout the same values, so
it was confusing to leave the old values there.