Show unread state.
Don't use gmail's read state. Keeping that in sync with maketime's
read state would require expensive syncing of everything in the inbox so
we can see if it's historyId has changed. Instead, have opening a thread
mark it as read in both maketime and gmail, but don't sync future unread
state from gmail (e.g. if you mark the thread unread in gmail, it will
stay read in maketime).