Add setting to disallow viewing messages in triage view.
Also remove Skim view and the concept of Skimming. It is mostly
redundant with the new setting and the concept of skimming was an
experiement that didn't work out well in practice. It was too easy to
get stuck in a place where you would skim for a day or two where your'e
particularly busy and then never have the time for regular triage on
non-busy days due to the build up from the busy days.