Don't include built-in labels in the QueuesView.
This makes it so we don't have to special case disallowing deletion of them.
In theory we could disallow deletion, but even then we wouldn't want archive
to show up in the queues list. Although, unfiltered being available so that
the user can choose sort order for it wouldn't be terrible.