Performance opimtizations for the ThreadListView with many rows.
- Fix early return in ThreadRow.render_ to actually early return. It
wasn't handling dates right so would never early return.
- Make ThreadRowGroups persistent in the DOM so that we don't recalc the
whole world on every frame.
- Use IntersectionObserver to display none offscreen ThreadRowGroup
rows.

Performance is now O(n) where n is the number of rows in groups that
intersect the viewport. Before n was the total number of rows in
uncollapsed groups. We should eventually do followup work to use
IntersectionObserver to track offscreen rows in a group that intersects
the viewport, but observing each row is way to expensive, so we'd need
to do something fancier.