More look and feel
- Move the pending container to the top, put a box shadow on it, and
make it scrollable with a max-height. Chose this max-height value since
it works well for both desktop and mobile rendering for when it hits the
limit to clearly see that there is content clipped.
- Don't show the all done indication until we've first loaded the
threadlistview data so we don't have it flicker on initial page load
when there are actually threads.