Remove redundant focus moving code.
On triage we would mvoe the focus when putting threads in the
queue to be triaged, but since we no longer remove them here,
we can instead rely on handleRowsRemoved_ to do so from
renderFrame_. This avoids the focus moving to the next row while
the current one is still being rendered.

There's still an issue where resetState causes the checkbox for the
current row to disappear while it's still being rendered, but that's
not quite as terrible.