Rework login code to login to gapi first, then firebase.
This is what the angularfirebase documentation recommends and it seems to
simplify the code and maybe work better. Hopefully this solves the issues of
multiple logins the first time you sign up for maketime.

Also make it so that we handle 401 failures by immediately redirecting to the
login screen from Net.js rather than waiting for 3 failures, then reloading only
to redirect to the login screen.

Lastly, factor out the login code to make it more legible.