Fix message rendering.
Sometimes the headers for each message render, but the message itself is
missing. The guess of what's happening is that the reused
QuoteElidedMessage is getting reparented into a RenderedThread other
than the one currently being rendered. Not sure how this could happen,
but do a speculative workaround to see if it still reproduces.