Fix some focus issues.
- Default focusImpliesSelected to true so it doens't get incorrectly cleared sometimes.
It should only be cleared when the user explicitly unchecks a checkbox.
- Do updatePendingArea after handleRemovedRows since gathering the list of removed rows
depends on them not having been moved already by the prepend calls in updatePendingArea.
This makes it so we correctly move focus to the next itme.
- Have the single thread view rely on actionInProgress to go to the next thread. This
has the same performance roughly and avoids having a special case codepath for single
thread view.