Fix error on undo.
We would get an error that the row isn't in a group on undo, which
it isn't because we haven't put it back in yet. INstead move that logic
into renderFrame_ and remove the hack for the single thread view case
that was working around the same issue of the model not having the row
that we set to this.renderedRow_ in hanldeUndo.