A number of performance optimizations.
- Immediately render when collecting threads to triage so the actionInProgress bit
hides the rows ASAP.
- Fetch the hasLabel data before hasPriority since it gets sorted to the top and that
way we fetch the data for those threads first.
- Stop using requestIdleCallback. It seemed to be causing significant idle time on
the main thread and was generally not worth the complexity as we tried to do multipel
fetches in parallel. Go back to using TaskQueue with 3 in parallel, but add the ability
to cancel the in progress task queue to stop fetching old threads on snapshot updates.
- Always wait for the first snapshot to have loaded before rendering since those show up
at the top. This avoids some layout instability.