Fix stuck from Compose.
We would not early return in applyLabel_ for messages you sent yourself
unless they had a priority. But marking something stuck from composeview
goesn't give it a priority. Everything else in that toolbar does except
for due. But for due, we don't want to early return in applyLabels since
we want that thread to still be in the inbox.