Fix dequeuing of threads that are both blocked and due.
We early return if a threads due date has already expired so we don't
keep forcing retriage every day of overdue items. But if an overdue
item is marked blocked, then we should still force retriage when
the blocked date is expired. So only early return for the due key.