Fix iOS button menus.
If you start the button drag action on the description text at the bottom
of the button, then the pointer up action doesn't work and you get
stuck with a permanent orphaned menu. pointer-events:none fixes it
by just making it so that it's like that element isn't there, but
not really understanding why it's necessary in the first place still.